+++ src/htsp_server.c
@@ -29,9 +29,7 @@
 #include "streaming.h"
 #include "htsmsg_binary.h"
 #include "epg.h"
-#include "plumbing/tsfix.h"
 #include "imagecache.h"
-#include "descrambler.h"
 #include "descrambler/caid.h"
 #include "notify.h"
 #include "htsmsg_json.h"
